requirements.txt
requests
python-dotenv
beautifulsoup4
numpy
pyyaml
pandas
# dotenv